backend/requirements.txt
asgiref==3.7.1
backports.zoneinfo;python_version<"3.9"
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.6.2
tzdata==2023.3
python-decouple==3.1
psycopg2-binary
openai~=0.28.1
SQLAlchemy==2.0.23
annotated-types==0.6.0
anyio==3.7.1
dataclasses-json==0.6.1
exceptiongroup==1.1.3
greenlet==3.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.331
langsmith==0.0.60
marshmallow==3.20.1
mypy-extensions==1.0.0
pydantic==2.4.2
pydantic-core==2.10.1
python-dotenv==1.0.0
sniffio==1.3.0
tenacity==8.2.3
typing-inspect==0.9.0
regex==2023.10.3
tiktoken==0.5.1
beautifulsoup4==4.10.0